Split ScriptChunk and MglScriptHeader from mgl_script_extr to libsat for reusability
